Alternative to the use of changepoint
Detection of mean change with changepoint may not be adapted to continuous transitory changes as well as when autocorrelation occurs in fluctuations due to the influence of drift.

The alternative option fits a reference function of richness dynamics over time with a first exponential part and a second constant part. The limit of the two parts represents the time to reach equilibrium.